Enable kernel stack recycling

+++ kernel/src/task/mod.rs
@@ -54,7 +54,8 @@ lazy_static! {
             extern "C" {
                 fn _num_app();
             }
-            let num_app_ptr = _num_app as usize as *const usize;
+
+            let num_app_ptr = _num_app as *const usize;
             let num_app = num_app_ptr.read_volatile();
             let mut app_start: [usize; MAX_APP_NUM + 1] = [0; MAX_APP_NUM + 1];
             let app_start_raw: &[usize] =
